+++ Ethereal Artefacts/Ethereal Artefacts/Assets.xcassets/Stargazer’s Tea Set.imageset/Contents.json
