+++ Ethereal Artefacts/Ethereal Artefacts/Assets.xcassets/Stargazer’s Tea Set.imageset/Mask Group 3.png
