+++ Ethereal Artefacts/Ethereal Artefacts/Views/HomepageView.swift
@@ -1,0 +1,20 @@
+//
+//  HomepageView.swift
+//  Ethereal Artefacts
+//
+//  Created by Albert Khurshudyan on 7.05.23.
+//
+
+import SwiftUI
+
+struct HomepageView: View {
+    var body: some View {
+        CartButton()
+    }
+}
+
+struct HomepageView_Previews: PreviewProvider {
+    static var previews: some View {
+        HomepageView()
+    }
+}
